delete vk-admin
