+++ .wp-env.json
@@ -2,5 +2,10 @@
     "core": null,
     "plugins": [
         "."
-    ]
+	],
+	"config": {
+		"WP_DEBUG": true,
+		"SCRIPT_DEBUG": true,
+		"FS_METHOD": "direct"
+	}
 }
