--- inc/vk-admin/package/config.rb
@@ -1,9 +1,0 @@
-http_path = "/"
-css_dir = "css"
-sass_dir = "_scss"
-images_dir = "images"
-javascripts_dir = "js"
-# output_style = :compact
-output_style = :compressed
-line_comments = false
-# sass_options = {:debug_info => false}
